+++ streamparse/dsl/exceptions.py
@@ -1,0 +1,6 @@
+"""
+DSL exceptions
+"""
+
+class TopologyError(ValueError):
+    pass
