+++ streamparse/storm/component.py
@@ -3,7 +3,7 @@ Module to add streamparse-specific extensions to pystorm Component classes
 """
 import pystorm
 
-from ..dsl.component import ComponentSpecification
+from ..dsl.component import ComponentSpec
 
 
 class Component(pystorm.component.Component):
@@ -14,6 +14,6 @@ class Component(pystorm.component.Component):
 
     @classmethod
     def spec(cls, name=None, inputs=None, parallelism=None, config=None):
-        return ComponentSpecification(cls, name=name, inputs=inputs,
-                                      parallelism=parallelism,
-                                      config=config, outputs=cls.outputs)
+        return ComponentSpec(cls, name=name, inputs=inputs,
+                             parallelism=parallelism, config=config,
+                             outputs=cls.outputs)
