+++ test/streamparse/test_dsl.py
@@ -1,95 +1,68 @@
 import unittest
 import logging
 
-from pystorm.spout import Spout
-from pystorm.bolt import Bolt
-
-from streamparse.dsl.topology import Topology, Grouping, TopologyError
-from streamparse.dsl.spec import Spec
+from streamparse.dsl.exceptions import TopologyError
+from streamparse.dsl.stream import Grouping
+from streamparse.dsl.topology import Topology
+from streamparse.storm.bolt import Bolt
+from streamparse.storm.spout import Spout
 
 
 log = logging.getLogger(__name__)
 
 
 class WordSpout(Spout):
-    streams = ["word"]
+    outputs = ["word"]
 
 
 class WordCountBolt(Bolt):
-    streams = ["word", "count"]
+    outputs = ["word", "count"]
 
 
 class TopologyTests(unittest.TestCase):
     def test_basic_spec(self):
         class WordCount(Topology):
-            word_spout = Spec(
-                WordSpout,
-                parallelism=2,
-            )
-            word_bolt = Spec(
-                WordCountBolt,
-                source=word_spout,
-                group_on=Grouping.fields("word"),
-                parallelism=8,
-            )
+            word_spout = WordSpout.spec(parallelism=2)
+            word_bolt = WordCountBolt.spec(inputs={word_spout:
+                                                   Grouping.fields("word")},
+                                           parallelism=8)
 
         self.assertEqual(len(WordCount.specs), 2)
-        self.assertEqual(WordCount.word_bolt.sources[0], WordCount.word_spout)
-        self.assertEqual(WordCount.word_bolt.group_on, ['word'])
+        self.assertEqual(list(WordCount.word_bolt.inputs.keys())[0],
+                         WordCount.word_spout)
+        self.assertEqual(WordCount.word_bolt[WordCount.word_spout['default']],
+                         Grouping.fields('word'))
 
     def test_bolt_before_source(self):
         class WordCount(Topology):
-            word_bolt = Spec(
-                WordCountBolt,
-                source='word_spout',
-            )
-            word_spout = Spec(WordSpout)
-
-        self.assertEqual(WordCount.word_bolt.sources[0], WordCount.word_spout)
+            word_spout = WordSpout.spec()
+            word_bolt = WordCountBolt.spec(inputs=word_spout)
+        self.assertEqual(WordCount.word_bolt.inputs[0], WordCount.word_spout)
 
     def test_invalid_spout_source(self):
         with self.assertRaises(TypeError):
             class WordCount(Topology):
-                word_spout = Spec(
-                    WordSpout,
-                    source='unknown',
-                )
+                word_spout = WordSpout.spec(source='unknown')
 
     def test_invalid_bolt_source(self):
         with self.assertRaises(TopologyError):
             class WordCount(Topology):
-                word_bolt = Spec(
-                    WordCountBolt,
-                )
+                word_bolt = WordCountBolt.spec()
 
     def test_invalid_parallelism(self):
         with self.assertRaises(ValueError):
             class WordCount(Topology):
-                word_spout = Spec(
-                    WordSpout,
-                    parallelism=0,
-                )
+                word_spout = WordSpout.spec(parallelism=0)
 
     def test_invalid_bolt_group_on(self):
         with self.assertRaises(TopologyError):
             class WordCount(Topology):
-                word_spout = Spec(
-                    WordSpout,
-                )
-                word_bolt = Spec(
-                    WordCountBolt,
-                    source=word_spout,
-                    group_on=Grouping.fields("foo"),
-                )
+                word_spout = WordSpout.spec()
+                word_bolt = WordCountBolt.spec(inputs={word_spout:
+                                                       Grouping.fields("foo")})
 
     def test_duplicate_name(self):
         with self.assertRaises(TopologyError):
             class WordCount(Topology):
-                word = Spec(
-                    WordSpout,
-                )
-                word_ = Spec(
-                    WordCountBolt,
-                    name='word',
-                    source=word,
-                )
+                word = WordSpout.spec()
+                word_ = WordCountBolt.spec(name='word', inputs=[word])
